requirements.txt
asgiref==3.3.4
certifi==2019.3.9
cffi==1.14.5
chardet==3.0.4
cryptography==3.4.7
cycler==0.10.0
defusedxml==0.6.0
Django==2.2.14
django-cors-headers==3.7.0
django-debug-toolbar==1.10.1
django-filter==2.4.0
django-rest-knox==4.1.0
django-webpack-loader==1.1.0
djangorestframework==3.12.4
graphics.py==5.0.1.post1
idna==2.8
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.20.3
oauthlib==3.0.1
Pillow==8.3.0
pycodestyle==2.5.0
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.5
requests==2.21.0
requests-oauthlib==1.2.0
scipy==1.6.3
six==1.16.0
sqlparse==0.2.4
stripe==2.27.0
urllib3==1.24.2
